Fix bug changes in data input prop are not reflected
Fixed the bug outline in the issue #215 where the angular component
does not update the editor contents
on changed data property.

Implemented OnChanges interface that checks if the data
property has been changed.
If it has been, then we write a new value to the editor.

Lastly, added one more unit test case to assert the fix.